requirements.txt
alpaca-py>=0.30.0
pandas>=2.2.2
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
prometheus-client>=0.20.0
